CloudBot/requirements.txt
pyrebase4
pyTelegramBotAPI
python-dotenv
